refactor(export): extract documentation generators to exportDocs module
`exportService.ts` shrinks from 2480 → 1824 LoC (-656) by hoisting 7
documentation-generation methods to `backend/src/services/export/exportDocs.ts`.
Each is a pure string-building or `fs.writeFile` wrapper with no class
state — ideal extraction target.

Extracted (now standalone exported functions):
- `generateReadme` — top-level export README
- `generateMetricsGuide` — long markdown describing every per-polygon
  metric + the spheroid_invasive Disintegration Index appendix
- `generateAnnotationGuides` — orchestrator for per-format guides
- `generateCocoGuide`, `generateYoloGuide`, `generateJsonGuide` —
  per-format QUICK_SETUP.md
- `generateMainAnnotationGuide` — cross-format index README

Class shell now imports the three orchestration entry points
(`generateReadme`, `generateMetricsGuide`, `generateAnnotationGuides`)
and delegates. The internal `generateCoco/Yolo/Json/MainAnnotation`
helpers are called by `generateAnnotationGuides` via direct function
calls (no class indirection needed).

`ProjectWithImages` type is now exported from `exportService.ts` so the
extracted module can reference it.

Verification:
- Backend `npx tsc --noEmit` passes
- `exportService.test.ts` + `exportController.test.ts`: 41/41 tests pass
- Prettier formatted
